requirements.txt
Flask==3.0.0
gunicorn==21.2.0
pyTelegramBotAPI==4.14.0
requests==2.31.0
geopy==2.4.1
python-dotenv==1.0.0
Pillow==10.1.0
